Update composer.json
After testing, you received the following error message:
****************************************
Your requirements could not be resolved to an installable set of packages.

  Problem 1
    - Root composer.json requires qcubed / composer dev-qcubed4 -> satisfiable by qcubed / composer [dev-qcubed4].
    - qcubed / composer dev-qcubed4 requires composer-plugin-api ^ 1.0 -> found composer-plugin-api [2.0.0] but it does not match the constraint.
  Problem 2
    - qcubed / composer dev-qcubed4 requires composer-plugin-api ^ 1.0 -> found composer-plugin-api [2.0.0] but it does not match the constraint.
    - qcubed / application dev-master requires qcubed / composer dev-qcubed4 -> satisfiable by qcubed / composer [dev-qcubed4].
    - Root composer.json requires qcubed / application dev-master -> satisfiable by qcubed / application [dev-master].

You are using Composer 2, which some of your plugins seem to be incompatible with. Make sure you update your plugins or report a plugin issue to ask them to support Composer 2.
****************************************

We will not know before this fix has been made and tested. Then you can see if you need to rewrite something else.

So much information for you! The Composer version has been updated to 2.0.8. The dependencies could not be updated in the development due to the old version of composer.